Move error handling views into core app
The cfgov folder is for project files. View code better belongs in
a Django app, and core is the best fit.
